moved FieldSlicer to Diagnostics.jl since it's now also used by WindowedSpatialAverage
