+++ src/Diagnostics/Diagnostics.jl
@@ -17,6 +17,7 @@ using Oceananigans: AbstractDiagnostic
 include("nan_checker.jl")
 include("field_maximum.jl")
 include("cfl.jl")
+include("field_slicer.jl")
 include("windowed_spatial_average.jl")
 
 end
